Update p10844.cpp
unsigned long long && mod 1000000000

+++ p10844.cpp
@@ -8,12 +8,12 @@ int				main(void)
 	ios::sync_with_stdio(false);
 	cin.tie(0);
 
-	int		N, tmp1, tmp2;
-	vector<int>	nums(10);
+	unsigned long long		N, tmp1, tmp2;
+	vector<unsigned long long>	nums(10);
 
 	cin >> N;
 	fill(nums.begin() + 1, nums.end(), 1);
-	for (int i = 1; i < N; i++) {
+	for (unsigned long long i = 1; i < N; i++) {
 		for (int j = 0; j < 10; j++) {
 			tmp2 = nums[j];
 			if (j == 0) nums[j] = nums[j + 1];
@@ -26,7 +26,7 @@ int				main(void)
 			tmp1 = tmp2;
 		}
 	}
-	int		ans = 0;
+	unsigned long long		ans = 0;
 	for (int i = 0; i < 10; i++) ans += nums[i];
 	ans = ans % 1000000000;
 	cout << ans;
